backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
scikit-image==0.22.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
gunicorn==21.2.0
python-dotenv==1.0.0
cloudinary==1.36.0
